refactor: add personal artwork register schema
