+++ src/Router.tsx
@@ -39,8 +39,8 @@ import { MyQuestionsPage } from './pages/MyQuestionsPage';
 import { MyReviewPage } from './pages/MyReviewPage';
 import { NotFound } from './pages/NotFound';
 import { OnboardingPage } from './pages/onboarding';
+import { PersonalArtworksRegister } from './pages/personal-artworks-register';
 import { PersonalArtworkDetailPage } from './pages/PersonalArtworkDetailPage';
-import { PersonalArtworksRegister } from './pages/PersonalArtworksRegister';
 import { PolicyPage } from './pages/PolicyPage';
 import { ExhibitionRegisterComplete } from './pages/RegisterCompletePage';
 import { SearchPage } from './pages/SearchPage';
