+++ src/pages/personal-artworks-register/index.ts
@@ -1,0 +1,1 @@
+export { PersonalArtworksRegister } from './PersonalArtworksRegister';
